refactor: useContractData hook
