+++ apps/dapp/src/components/ssov-new/Tables/StrikesChain/index.tsx
@@ -3,7 +3,7 @@ import StrikesTable from './StrikeTable';
 
 const StrikesChain = ({ market }: { market: string }) => {
   return (
-    <div className="space-y-2 bg-cod-gray rounded-lg pt-3">
+    <div className="bg-cod-gray rounded-lg pt-3">
       <div className="relative h-12 mx-3">
         <FilterPanel market={market} />
       </div>
