+++ apps/dapp/src/utils/date/getTimeToExpirationInYears.ts
@@ -1,11 +1,11 @@
 const SECONDS_IN_A_YEAR = 31536000;
 
-const getTimeToExpirationInYears = (expiry: number) => {
+const getTimeToExpirationInYears = (expiryInSeconds: number) => {
   const now = Date.now() / 1000; // Convert ms to s
 
-  if (now > expiry) return 0;
+  if (now > expiryInSeconds) return 0;
 
-  const timeToExpirationInSeconds = expiry - now;
+  const timeToExpirationInSeconds = expiryInSeconds - now;
 
   return timeToExpirationInSeconds / SECONDS_IN_A_YEAR;
 };
